HACKLAB HACKTHON PWNSTARS PROJECT FOR CRIME PREDICTION
HACKLAB HACKTHON PWNSTARS PROJECT FOR CRIME PREDICTION
Introduction
San Francisco was infamous for housing some of the world's most notorious criminals on the inescapable island of Alcatraz. Today, the city is known more for its tech scene than its criminal past. From Sunset to SOMA, and Marina to Excelsior, this project analyzes 12 years of crime reports from across all of San Francisco's neighborhoods to create a model that predicts the category of crime that occurred, given time and location.

Definition
Project Overview
Crime is a social phenomenon as old as societies themselves, and although there will never be a free from crime society - just because it would need everyone in that society to think and act in the same way - societies always look for a way to minimize it and prevent it. In the modern United States history, crime rates increased after World War II, peaking from the 1970s to the early 1990s. Violent crime nearly quadrupled between 1960 and its peak in 1991. Property crime more than doubled over the same period. Since the 1990s, however, crime in the United States has declined steadily. Until recently crime prevention was studied based on strict behavioral and social methods, but the recent developments in Data Analysis have allowed a more quantitative approach in the subject. We will explore a dataset of nearly 12 years of crime reports from across all of San Francisco's neighborhoods, and we will create a model that predicts the category of crime that occurred, given the time and location.

Problem Statement
To examine the specific problem, we will apply a full Data Science life cycle composed of the following steps:

Data Wrangling to audit the quality of the data and perform all the necessary actions to clean the dataset.
Data Exploration for understanding the variables and create intuition on the data.
Feature Engineering to create additional variables from the existing.
Data Normalization and Data Transformation for preparing the dataset for the learning algorithms (if needed).
Training / Testing data creation to evaluate the performance of our models and fine-tune their hyperparameters.
Model selection and evaluation. This will be the final goal; creating a model that predicts the probability of each type of crime based on the location and the date.